sundayleague/requirements.txt
pytest
pycodestyle
mysql-connector-python